Move POSIX logic to a separate static reference, to avoid initialing on Windows
